Description for the SystemService interface was added.
