+++ Back-End/src/main/java/com/c1646njava/tuvivienda/ServletInitializer.java
@@ -9,5 +9,4 @@ public class ServletInitializer extends SpringBootServletInitializer {
 	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
 		return application.sources(TuviviendaApplication.class);
 	}
-
 }
