+++ Back-End/src/main/java/com/c1646njava/tuvivienda/controllers/SystemController.java
@@ -1,0 +1,4 @@
+package com.c1646njava.tuvivienda.controllers;
+
+public class SystemController {
+}
