+++ Back-End/src/main/java/com/c1646njava/tuvivienda/services/SystemServiceImpl.java
@@ -1,0 +1,4 @@
+package com.c1646njava.tuvivienda.services;
+
+public class SystemServiceImpl {
+}
