--- Back-End/src/main/java/com/c1646njava/tuvivienda/services/abstraction/SystemService.java
@@ -1,20 +1,0 @@
-package com.c1646njava.tuvivienda.services.abstraction;
-
-import java.util.List;
-
-import com.c1646njava.tuvivienda.models.post.Post;
-
-public interface SystemService {
-    
-    Long registerUser(String name, String Password,String email);
-    Boolean loginUser(String email, String Password);
-    Boolean logoutUser();
-    Boolean checkSession();
-    List<Post> searchByLocation(String address);
-    List<Post> searchByType(String type);
-    List<Post> searchByBedrooms(Integer bedrooms);
-    List<Post> searchByPrice(Long priceLow,Long PriceHigh);
-     
-
-
-}
