--- Back-End/src/main/java/com/c1646njava/tuvivienda/services/abstraction/UserService.java
@@ -1,7 +1,0 @@
-package com.c1646njava.tuvivienda.services.abstraction;
-
-public interface UserService {
-    
-    
-
-}
